refactor(studio): improve code reuse and make it generally cleaner
