+++ apps/studio/lib/entities/lens/index.ts
@@ -1,0 +1,16 @@
+export type {
+  BuildHostQueryOptions,
+  LensArgs,
+  LensDefinition,
+  LensResultKind,
+  ResolvedLensView,
+  TokenAddressLensArgs,
+} from "./model/types";
+export {
+  STUDIO_LENS_CATALOG,
+  getLensDefinition,
+  isStudioSupportedLens,
+  type AnyLensDefinition,
+} from "./model/catalog";
+export { ERC20_TRANSFER_LENS } from "./model/erc20-transfers";
+export { ERC20_ACCOUNT_BALANCES_LENS } from "./model/erc20-account-balances";
