+++ apps/studio/lib/entities/view/index.ts
@@ -1,0 +1,12 @@
+export type {
+  DeployProgressStatus,
+  DeployStatus,
+  LensQueryPage,
+  StoredDeployedView,
+  StoredDeployedViewSource,
+} from "./model/types";
+export { STUDIO_QUERY_LIMIT } from "./model/types";
+export { StoredViewsProvider, useStoredViews } from "./model/use-stored-views";
+export { useDeployLens } from "./model/use-deploy-lens";
+export { ViewValidationError } from "./model/view-validation-error";
+export { StoredViewsPanel } from "./ui/stored-views-panel";
