--- apps/studio/lib/pages/studio/header.tsx
@@ -1,17 +1,0 @@
-"use client";
-
-import { ConnectDialog } from "@/shared/ui/connect-dialog";
-import ShinzoLogo from "./shinzo-logo.svg";
-
-export function Header() {
-  return (
-    <header className="flex items-center justify-between border-b border-szo-border px-6 py-4">
-      <ShinzoLogo
-        className="h-8 w-auto shrink-0"
-        aria-label="Shinzo"
-        role="img"
-      />
-      <ConnectDialog />
-    </header>
-  );
-}
