+++ apps/studio/lib/shared/consts/query.ts
@@ -4,8 +4,8 @@ import {
   isServer,
 } from "@tanstack/react-query";
 
-function makeQueryClient() {
-  return new QueryClient({
+const makeQueryClient = () =>
+  new QueryClient({
     defaultOptions: {
       queries: {
         staleTime: 60 * 1000,
@@ -17,15 +17,13 @@ function makeQueryClient() {
       },
     },
   });
-}
 
 let browserQueryClient: QueryClient | undefined = undefined;
 
-export function getQueryClient() {
+export const getQueryClient = () => {
   if (isServer) {
     return makeQueryClient();
-  } else {
-    if (!browserQueryClient) browserQueryClient = makeQueryClient();
-    return browserQueryClient;
   }
-}
+  if (!browserQueryClient) browserQueryClient = makeQueryClient();
+  return browserQueryClient;
+};
